add ability to edit course only for owner

+++ routes/courses.js
@@ -11,7 +11,12 @@ router.get('/', async (req, res) => {
       .populate('userId', 'name email')
       .select('title price imageURL')
 
-    res.render('courses', { title: 'Courses', isOnCoursesPage: true, courses })
+    res.render('courses', {
+      title: 'Courses',
+      isOnCoursesPage: true,
+      userId: req.user ? req.user._id : null,
+      courses
+    })
   } catch (error) {
     console.error(error)
   }
@@ -31,16 +36,18 @@ router.get('/:id', async (req, res) => {
 })
 
 router.get('/edit/:id', authMiddleware, async (req, res) => {
-  if (!req.query.allow) {
-    return res.redirect('/')
-  }
-
   try {
     const course = await Course.findById(req.params.id)
-    res.render('course-edit', {
-      title: `Edit ${course.title}`,
-      course
-    })
+    const isCurrentUserCourse = course.userId.toString() === req.user._id.toString()
+
+    if (isCurrentUserCourse) {
+      return res.render('course-edit', {
+        title: `Edit ${course.title}`,
+        course
+      })
+    } else {
+      return res.redirect('/courses')
+    }
   } catch (error) {
     console.error(error)
   }
@@ -48,8 +55,15 @@ router.get('/edit/:id', authMiddleware, async (req, res) => {
 
 router.post('/edit/:id', authMiddleware, async (req, res) => {
   try {
-    await Course.findByIdAndUpdate(req.params.id, req.body)
-    res.redirect('/courses')
+    const course = await Course.findById(req.params.id)
+    const isCurrentUserCourse = course.userId.toString() === req.user._id.toString()
+
+    if (isCurrentUserCourse) {
+      await Course.findByIdAndUpdate(req.params.id, req.body)
+      return res.redirect('/courses')
+    } else {
+      return res.redirect('/courses')
+    }
   } catch (error) {
     console.error(error)
   }
@@ -57,7 +71,7 @@ router.post('/edit/:id', authMiddleware, async (req, res) => {
 
 router.post('/delete/:id', authMiddleware, async (req, res) => {
   try {
-    await Course.deleteOne({ _id: req.params.id })
+    await Course.deleteOne({ _id: req.params.id, userId: req.user._id })
     res.redirect('/courses')
   } catch (error) {
     console.error(error)
